requirements.txt
streamlit
opencv-python
pillow
easyocr
pyzbar
matplotlib
numpy
pandas